Place class was created
