+++ elephant/src/main/java/ru/anutakay/Box.java
@@ -2,13 +2,12 @@ package ru.anutakay;
 
 import ru.anutakay.exception.*;
 
-public interface Box {
+public interface Box extends Place {
 
     void put(Freezable object) throws BasicException;
 
     Freezable get() throws BasicException;
 
-    boolean isFits(Freezable object);
-
     boolean isFull() throws BasicException;
+
 }
